Fixed issue in filter leading to sanitizer errors.
vtkDataSetSurfaceFilter had some funky code designed to optimize
(minimize) allocation of small objects. This required apparently
writing beyond the limits of an array. It was not a bug but made
tools like the clang sanitizer think that it was. Fixed it while
still maintaining the allocation optimization, at the cost of a
bit more memory usage.

Change-Id: I3ef7051e85601a4440f02cb4dfe2e15227a06e87