Changes to allow for multiblocks and seeds to change over time steps.
By default this class assumed that multiblock dataset inputs
and seeds would not change over time. This is a bad assumption.
Also, there were variables to turn this behavior off but the
setter methods were never created. Now, the more general behaviour
is to assume that multiblock data sets and seeds may change over
time and to provide the setter methods to change to the more
efficient calculations based on static information.

Change-Id: Ib4cf5ef818f314708d1aac5e2247a66b59fe7654