COMP: Test driver array bounds error
Some compilers report "array subscript is above array bounds" while
compiling vtkRenderingMatplotlibCxxTests. This occurs when
Module_vtkRenderingMatplotlib:BOOL=ON but Matplotlib is not found or
VTK_GHOSTSCRIPT_EXECUTABLE is not defined. In this case, there are no
tests defined, but the test driver is still built.

Rather than change CMake, this patch adds an informative dummy test.

Change-Id: I3f1d6ed2453d33d113f165fc9552f18d2cc4906f