Added "sticky" axes mode to vtkCubeAxesActor
This change adds a mode to vtkCubeAxesActor that keeps the axes in the
viewport by changing the boundaries of the axes to keep them all
within the viewport. It works by computing the largest sphere that
will fit in the view frustum at the depth of the original bounding
box's center, then computes a cube that fits within that sphere.  The
final bounding box is the intersection of the original bounding box
and the cube that fits within the sphere. If this intersection is
empty, the bounds are invalidated so that no axes are displayed.

The sticky axes mode also has a centering option that controls whether
the truncated bounding box stays centered in the viewport or follows
the center of the original bounding box along the longer of the two
viewport dimensions. Whether this option is enabled or not, the axes
are restricted to being within the viewport.

Change-Id: I6dc3cd4127d2e4da8f2580fd2242ffa51270629a