Fixed error in vtkFiltersCoreCxx-TestClipPolyData test
It was logging:
vtkPlanes: Please define points and/or normals!

The cause was a vtkPlanes object that was created but not usefully initialized.
Right after the log, DBL_MAX was returned and later cast to float, which is
too small to hold DBL_MAX and clang's ubsan flagged that error.

I added any old initialisation of the vtkPlanes, which is sufficient
since the test doesn't look too closely at the results of all its computations.

Thanks to Will at the hackathon for helping me figure this out!

Change-Id: I08f2863a718cb82fa54ecc7e2f1f3337f43b684d