Add a delay between process launches.
It so happens that the client process was launched before the server causing
deadlocks in some cases. Fixed that by adding a small delay. In reality, we may
need to logic to vtkTestDriver to make it more featured and parse outputs from
processes to handle this correctly. We can do that in next stage.

Change-Id: Ief6989c50e816bc8b6b7f8a780fd31b578079f4c