Fix and enhance the box plot and chart
Fix the problem with the position of Y axis: long labels were trunked.
This position was fixed, it is not flexible depending the size of the labels.

Add new functions to the chart API to allow manipulation of series
by id (vs by name).
Add support for manual plot labels setting.
Enhance the boxplot displacement, it now correctly follow the mouse
without ugly jumps.

Change-Id: I400cd064616f7b93415e17db9749d709ea2f7345