Allow custom TestTM3DLightComponents timeout
Generalize the timeout increase in commit 15605b33 (Increased the
timeout slightly to allow passing on slow machines, 2013-03-03) to allow
configuration of the timeout in the local cache.  This will allow it to
be increased on the few machines where we need to do so (typically for
software-only GL) without increasing it everywhere.

Change-Id: I37be08ab9fd6e2b2cab1239b55d0874aed577bf3