Add SetSlabModeToSum() to vtkImageResliceMapper.
This change allows vtkImageResliceMapper to produce images that look
like digitally reconstructed radiographs.  It required a new property
to be added to vtkImageReslice, called ScalarScale, so that the output
intensity can be appropriately scaled to the slab thickness, which is
specified in physical units of distance (typically millimeters).  An
existing test was modified in order to test this new feature.

A ScalarShift property was also added for symmetry with ScalarScale,
plus an OutputScalarType property so that vtkImageReslice can output a
different scalar type than its input (useful for avoiding overflow with
SetSlabModeToSum()).

This change also fixes a performance issue that was introduced in August
2011 when the SetInterpolator() method was added: that change caused the
scalar conversion to always clamp to the limits of the output type, even
when clamping was unecessary.  Now, clamping is only done when needed.

Change-Id: I5e55db509e8ec537632a7a7e063a3bc0c50d9d20