break vtkRCalculatorFilter's input/output symmetry
Before this commit, vtkRCalculatorFilter produced the same type
of output as its input.  This is no longer the case.  For example,
you can now run an R script that takes one table as input
and produces two trees as output.  A test is included for this
new functionality.

This commit also makes vtkRAdapter::RToVTKTree() more robust.
Previously, we assumed that the elements in the R Tree would
appear in a certain order, but apparently that is not guaranteed
to be true.  We now parse these elements in whatever order they
appear.  While I was modifying this function, I also cleaned it up
so it should now adhere to VTK's coding style.

Change-Id: I988361ee8e6788f6f51e0c930fc52df50d1b2912