ENH: Provide PlaybackFile for recorded interaction events
Many of the Widget tests supply in-code event logs. This patch adds
an additional option to vtkTesting::InteractorEventLoop. The new
option, --PlaybackFile, supplies a file name for the
InteractionRecorder. This allows tests to modify the playback
events without changing the source code for the test. This separates
changes to the playback events from changes to the test code.

A couple of tests have been modified to illustrate the usage of the
InteractorEventLoop.

In the future, the current widget tests should be changed to use the
InteractionEventLoop rather than ifdef'ed code for playing and
recording.

Change-Id: I6aa2afd7c161fe82646fc1df7d326ca8a3249a2f