Set MACOSX_RPATH property default consistently across CMake versions
The MACOSX_RPATH target property was introduced in CMake 2.8.12, but we
currently allow CMake 2.8.8.  Set CMAKE_MACOSX_RPATH to 0 by default
so that we get consistent behavior across all versions of CMake.
Leave a comment explaining when the default should be updated.

Now that we set a MACOSX_RPATH default explicitly, we no longer need to
set CMake Policy CMP0042 as was added in commit 2ad9f474 (Set CMake
Policies CMP0025 and CMP0042 as necessary, 2014-03-11), so remove it.

Change-Id: I8af24a728379176b9e7ab933206eaa956e1c60f2