Limiting number of ModifiedEvent when calling vtkTransform::SetMatrix
When providing one matrix to vtkTransform, there were 2 calls to Modified()
(one due to the internal call to Identity and one to concatenate the provided
matrix). So too many ModifiedEvent were sent, which could lead in some
particular set up to too many updates, and slow down execution.

Here a direct call to this->Contatenation->Identity() avoid a non necessary
call to Modified(), and so the number of ModifiedEvent is now limited to 1 when
setting one matrix.

Change-Id: I04c96776d168c75e5f29ee326ff7ed84458a40b3