fix potential segmentation fault
If one table and one tree were requested as output from
vtkRCalculatorFilter, its output would be a vtkTable or vtkTree.
It actually needs to be a vtkMultiPieceDataSet to contain both
objects though, and this would result in a segmentation fault
shortly thereafter.  This commit fixes this degenerate case
and adds a test for it.

Change-Id: I53e468bba2e2d336dc7a4a7f9fb53fdcb64f98b6