Roughly mapping the colors/scalars, updates to shaders
Attempt to map the colors/scalars, and adjust the shader according
to whether color is a uniform or an attribute. Added some logic to
deal with building normals as the painters do, but error out for
now. This is also not even close to dealing with cell colors and/or
properties.

Change-Id: If584570b71dca4db3e9fd6c4ad95af58c00dea32