Early attempt at getting the 3D glyph mapper working
It is very boring right now as it skips the color mapping, it looks
like it would be easiest to create a color mapping class/function
that duplicates what the scalars to color painter did. It looks like
TestGlyph3DMapper is pretty close once color is passed through.

I think the API on the VBOPolyDataMapper will need work too, but it
effectively passed raw colors before using fixed pipeline color calls.
We will need to cache the transforms to make this in any way efficient
as the previous logic relied on display lists caching this stuff more
implicitly.

Change-Id: I94bb5f352eb235341bb24b0676c00d361aa40325