Support faster configuration times by checking dependencies at runtime.
Now all web-based tests have the option to check some dependencies at
runtime and raise a "DependencyError" in case some requirement is not
met.  For the time being, these kinds of errors are caught by the
testing framework, and the test is allowed to pass.  However, in the
future when we have access to the new CMake feature, "SKIP_RETURN_CODE",
we can use it in place of passing the tests to indicate that the test
was not run because of unmet dependencies.

Change-Id: I1482d47f423ec68bdbcd27cb857829422fe65368