Very judicious use of sed, vtkOpenGL2* -> vtkOpenGL*
This now gives us a vtkRenderingOpenGL2 that has most of the same
classes as vtkRenderingOpenGL - the two will clobber each other's
headers, symbols etc and so can never be built/linked at the same
time.

Unfortunately a lot of classes SafeDownCast or dynamic_cast to the
derived OpenGL forms of the classes in vtkRenderingCore, and so
using a different class name presented significant issues despite
the fact that these classes are (in an ideal world) implementations
of the API in vtkRenderingCore for OpenGL 2.1 and above.

Change-Id: Ie8c6234621e550d3fd67da732d2cf4839290d665