Provide a method to set the sample spacing for slab modes.
The slab modes of vtkImageReslice allow for "thick slab" views, where
several slices are combined, for example, to create a maximum intensity
projection.  Previously the spacing between these slices was always set
to the output slice spacing, which was fine if the user only wanted to
create a single "thick slab", but less useful if the user wanted to create
a series of "thick slabs".  This patch adds a new method, called
SetSlabSpacingFraction, so that the user can first set the OutputSpacing
to set the spacing between the output slabs, and can then set the
SlabSpacingFraction to set the spacing between the slices that are
combined to make those slabs.

Change-Id: Ib70e92da9b1e27ba59a7a293d6e19b51ac25c9bb