Don't exclude the module from wrapping - initialization
I am not sure if this is the correct solution, but it matches what
we do for vtkRenderingOpenGL, etc. There is little point in us
wrapping these classes, but we want the relevant implementation
module to be linked/initialized by the wrapped language in order
to be able to render charts correctly using a concrete
implementation.

Change-Id: I2402b3578f5a89b10dad2619c955e900c7ab822b