PERF: Removed unnecessary garbage collection.
Information iterators increment/decrement the reference of their
information object by default. This leads to lots of garbage
collection passes, which when not necessary cause performance
degradation. Used a weak reference instead. In these cases, we
know for sure that someone else has a reference to the information
objects (the executive).

Change-Id: I9bc0f6a09b96c726b5d2a3bb7e78f3f7c67152b2