BUG: Make selections more permanent, linked to object
The bar charts selection was quite fragile, being attached to objects
that are deleted and recreated on any given update to the bar object.
This adds the selection directly to the selection array, and instead
of changing the color overpaints a bar blending with the original
colors, making it easier to see multiple different colored bars that
are selected.

Change-Id: I84b1128c26af5743af9868adc72e0da038ea775c