Cleanup MPI controller properly.
If the python script run by pvtkpython called sys.exit() explicitly, the
MPI controller cleanup logic wouldn't kick in and the controller would
leak causing test failures. Adding a Py_AtExit() handler ensures that
the controller is cleaned up when the Python interpreter is finalized,
one way or the other.

Change-Id: I5433fcc111b20b3f7f9809d4d8f61c1128b216e2