put array reuse in
xdmf arrays are shallow copied into VTK and deleted en masse at
the reader level instead of being deep copied and deleted for
every VTK user of them. With xincluded arrays especially, this is
a huge savings in terms of time and memory.
We use an LRU cache to delete only those arrays which were
not used in the previous frame and keep the rest around
because they are likely to be reused.

Change-Id: Ica75069db7ca1357f5a343c3e201ebe1428e1f0e