Fixed error in appending cell data
Arrays allocated for the appended cell data were sized to the number
of output points instead of the number of cells. In addition, the
global index array was being used to look up point indices when
copying cell data instead of cell indices.

Modified test to have a different number of points and cells in the
input data sets. Also, test arrays in inputs have random values to
ensure correct ordering in output.

Change-Id: Ia6c80bcc5f26dbf2b928633b8873b84a385cacf4