Fix handling of empty selection on ParallelCoordinates.
Same as ac920d1f4a, except this time for vtkChartParallelCoordinates.
The fix exposed a problem with vtkPlotParallelCoordinates where when the
selection was set to NULL, it was never updated after that. This was
incorrect (since for a similar condition in vtkPlotPoints, we create a
new vtkIdTypeArray, when none is present). Fixed that following the
pattern in vtkPlotPoints.

Change-Id: I77cd9d3c6f9f0a616261439fd2db6b2991147d55