Fix vtkAbstractPolygonalHandleRepresentation3D visibility
When the default position of the camera is not looking at 0,0,0 (default
handle position) the handle representation was never displayed.
Its bounds were culled by the renderer which prevented the handle actor
from being rendered which prevented the representation from calling
BuildRepresentation() which prevented the actor to have its correct
position set.

Change-Id: I36bffca7fef86c1493ff6869522fb89824390d06