Refactoring CMake code for Python modules.
Adding some infrastructure to properly handle vtk-modules that are
simply Python packages or modules.

This change adds API in vtkModuleMacrosPython to manage all the building
and installing of Python modules and packages minimizing the confusion.

Past attempts fix six.py, vtk/web etc. failed to distinguish cleanly
between a Python package and a module which was resulting in directories
being "unwrapped" on install (e.g. with autobahn) or py files getting
put in subdirectories (six.py).

This also manages installing on "vtk/web" package under the "vtk"
package cleanly. For that we needed to move the *.py files in Web/Python
into a subdirectory reflecting the targetting package structure.

Change-Id: Ic251c34418b07838e5bb483c2b02b12004f17632