Exit the Cocoa event loop when the window closes.
This is to match the behavior on other operating systems.  If the
native window was originally created by VTK (and not merely attached
to VTK), then when the user closes it, the interactor's Start()
method returns.

Without this behavior, basic interactor-based VTK scripts and programs
will simply hang when the window is closed.

Change-Id: I900acda4dc2f98999615be7a3715716c64bc49b2