vtkPythonAlgorithm: print out all the error information
Unfortunately, Python has this habit of using the name of the
exception's class for critical information and the stringification of
the exception itself is useless on its own (e.g., AttributeError's
string is just the name of the attribute that is missing). To get better
error information, use traceback.format_exception if it is available.

Change-Id: I7fcbd9a12e505b2e3422c31dcf52055d98144a64