Add support for OSMesa builds of OpenGL2
This commit makes some changes in Mesa's logic so that it can use
OSMesa's OpenGL context, and load extensions in OSMesa. This works
on Linux, but will need more work to get it working on Windows
and/or Mac OS X. Pulled out the OpenGL Mesa logic so that both GLEW
and vtkRenderingOpenGL2 can use it (and ensure it is consistent).

Change-Id: Ibdf6715f950929949f5055b3ddbd38b09d9b84d6