Add support for OpenGL32 instancing in GlyphMapper
Currently disabled as no 3.2 contexts are requested, but this patch
adds support on an OpenGL 3.2 context to use instancing for rendering
glyphs. This required adding some additional capabilities to the VAOs
so that they can handle matrices and divisors. This patch shows slightly
better performance on a 2.1 context and about a 100% improvement on
a 3.2 context. Picking on the 3.2 context path still needs some minor
work.

Change-Id: Ifb59184e9e3184b664fb642c5ffa73f4f268dd51