Fix compile warning in TextureObject
Texture object method was requiring the width and height of the
framebuffer as arguments.  It was doing error testing on these
arguments but there is no reason to.  Copying a texture where part of
the texture goes outside the framebuffer is fine. That can happen
and is not a problem. The extra texels just get discarded. This also
saves having to know the width and height of the frame buffer
to make this call.

Change-Id: I96cafda2ac900f1ecd247e6678e1e68cbccf033c