Expand support for the OpenGL2 backend
This change makes more modules dynamically switch between the
rendering backends if they have no OpenGL code, or introduces some
ifdefs in the case of RenderingParallel (which may not be functional
at this stage). Added RenderingFreeTypeOpenGL2 which will also need
some additional work before it is properly integrated with the new
rendering backend, but it surprisingly seems to render on the
desktop.

This commit contains everything needed to get a minimal ParaView
build using the new rendering backend for polygonal rendering.

Change-Id: If8676d41d263b046a54060965fc760b5d4bd64a1