14999: Python Tk widgets fail to load on Tcl 8.6.
The tcl 8.6 "info loaded" command returns a nested list of loaded
tcl modules.  In python, these are returned as nested tuples (if
there is at least one), or as an empty string (if there are none).
Both situations are now handled, as well as the tcl 8.5 behaviour
in which the results are always returned as a string.

This supersedes commit 9f4afe40, which returned a false negative
because it mistakenly assumed that "info loaded" returned a flat list.

Tested with tcl 8.5 and tcl 8.6 with python 2.7 on linux.

Change-Id: Ic9aa3b59fb141bd0bcf5e8f12106173fb1fc4c5b