Remove the deprecated PYTHON_INCLUDE_PATH variable.
This variable has been deprecated for years and its presence can
lead to extreme frustration.  If it is set in the cache, then it
is used as the PYTHON_INCLUDE_DIR.  That means that if someone
removes PYTHON_INCLUDE_DIR from their cache (e.g. while trying to
debug their build), then cmake will set PYTHON_INCLUDE_DIR from
the internally cached PYTHON_INCLUDE_PATH instead of searching the
system for Python.h like it should.

Also fix the variable names in the documentation for this macro,
which were different from the names of the variables that were actually
used.

Change-Id: I55d54b9d05085118c97a2e1d417e8e6b31289f8a