A first cut at a CompositePolyDataMapper2
So this is a very different use case from what we have done so far.
In many ways this is the lots of independent actors case with the
exception that they are not fully independent. There were a number
of CPU botlenecks hindering this classes performance:

  stl maps and strings
  recomputing shaders/uniforms
  GetInput/SetInput (touching the pipeline is bad)
  Default Culler

There are a couple open issues with this change. I'm
not thrilled with the CurrentInput approach in PolyDataMapper
but SetInput and GetInput really are expensive. Second there is
an issue if a bunch of datasets share a single
large pointset. Right now we duplicate the pointset VBO for each
dataset. Will need a VBO cache at some point to allow actors/etc
to share a pointset or have the VBO building only populate the
VBO with the points actually used by the cells (or both).

This patch also contains a minor fix to the VAO missing an
initialization for ForceEmulation.

Change-Id: If2cf9531a26838b0fd8d47dafd1dcff6d4133c4b