Make streaming optional for the 3D image mappers.
Previously, the default behavior of the 3D image mappers was to stream
the data through the pipeline, meaning that only the part of the input
image that was required to display one (possibly oblique) slice on the
screen was updated.  However, this resulted in the UPDATE_EXTENT changing
whenever a new slice was displayed, triggerring a pipeline update.

The new default behaviour is to set the UPDATE_EXTENT to the WHOLE_EXTENT,
so that the whole extent is updated on the first render, but no pipeline
execution occurs on subsequent renders (e.g. while slicing through the data)
unless the pipeline is modified.

The StreamingOn() method can be used to revert to the old behavior.

Change-Id: I76b31a2ae4b329885d755c2947b62f9a9474b88a