Get render passes working in OpenGL2
Add in a bunch of render pass classes that were previously removed.
At the same time get the GaussianBlurPass and SobelGradientMagnitude
classes to work with the new backend.
This code now exercises a few methods that previously were not
previously exercised. As such I made some significant changes to those
methods to make them fit in better with the new architecture.

In the process of testing I did find a calculation error in the old
code related to cmoputing camera angles. I fixed this issue in both the
old and new rendering backend.

The tests for these passes were modified to not use the DepthPeelingPass
as that pass has not been converted yet.

Change-Id: Ie40231385291aa06848034460817b171e60f7a53