Simplify and inline much of the logic in vtkStructuredData.
The dataDescription argument wasn't necessary for most of the computations
and, at best, saved a few integer operation at the cost of extra analysis
logic and branching.

Moved the workhorse functions to the header so they can be inlined.

For now, I've just documented that the dataDescription arguments aren't
used. Might be worth removing them from the API at some point.

Change-Id: Iea23e2127eba14614455c610df558b7753196b36