BUG: Some types of files are not closed after processing
The file is not closed after multi-page or tiled images are
processed. This results in unclosed file descriptors. Some systems
have limits on the number of open file descriptors. If a large number
of files are processed, subsequent open's may fail even if the file
exists.

Added a test to check if multiple reads do not fail.

Change-Id: I8e97496aae32bc8249b5b26a6101c81d6829ed1b