BUG #15046, BUG #15020: Fix issues with OpenGL 1.2.
Believe it or not, we have use-cases where people use OpenGL 1.2! Soon
we won't support that. But until then, we should continue to support
it. Based on discusson on BUG #15020, making the missing FBO warning to
be shown only in debug builds. Also use GL_ARB_multitexture explicitly
when GL_VERSION_1_3 is not supported.

Change-Id: Id8ec2e902536fb3165e03bc05c6a9bd1fcbc03e8