Add an enum type to the python wrappers.
Any enum types that are encountered will automatically be wrapped
as integral types.  This is true for enum types defined in the global
namespace, and enum types defined within a wrapped class.  Other
namespaces are not yet supported, because the python wrappers do not
yet wrap C++ namespaces.

When a class method has an enum parameter, and if the enum is a member
of the class, then that method will perform type checking on the enum.
Class methods that require enums that are not members of the class will
not be wrapped (in future versions of the wrappers, they might).

Change-Id: If38391de8a1a66eb255ddf83a72916af6d81dc78