Fix lighting to use Blinn-Phong specular model
A few fixes and cleanups. Use the Blinn-Phong specular lighting
model in the LightKit and Positional shaders. Old OpenGL uses this
model and it is a decent model. Remind myself that the camera is
positioned at 0,0,0 in view coordinates, had to fix a couple places
that were making different assumptions.

Minor changes to TestVBOPLYMapper to make it easier to use
when testing out changes.

Minor warning and documentation fixes as well.

Change-Id: I8bf108ec18fef2f0a2f83ada8cdf7663086036be