Improve picking support in OpenGL2
Point association picking was not implemented this patch
adds that in. Also cell picking did not work when the input
contained quads or higher vertex count polygons.  To address
these issues the picking process needs to give the mapper an
opporunity to map the index from the frame buffer back into
a point id or cell id. Added amvirtual method in prop to
do that and an implementation in OpenGLActor and
OpenGLPolyDataMapper to properly map gl_PrimativeId in
the correct point or cell id.

Point picking was not tested in VTK so I added a test for
that as well.

Change-Id: If276e80ab3966646dd9f6f95fbc7734c006f000c