Add in support for OpenGL ES 3.0
Make the build process use the opengl ES 3.0 header files if
selected and modify the backend to enable more features when
using ES 3.0. VAOs, FBOs, and instanced draw calls are enabled.

There are still some changes to be made for textures probably and
as with ES 2.0 only some modules have been tested/converted.

Change-Id: If244dc5bc9295a86f98b032b3a21995336a8fb05