Fix bugs in vtkStructuredImplicitConnectivity.
Removed some asserts in the registration functions, since often a
plane or line is passed in place of a grid due to partitioning
artifacts.

Removed a value from the communication stream that was not being read
and unsynchronizing the read/write process.

This seemed to assume that the VOI would pass through every process's
partition. Fixed that.

Also patched a possible memory leak during grid registration.

Change-Id: Ia16601c672225fc3a0c1a11480409394c3fd9e09