Fix some release graphics resource issues
ParaView IceT does a WindowRemap when using tile display. That
means everythgin has to work with a new window/context so it
identifies a bunch of release graphics resource issues. This
patch fixes a bunch of them. Mostly minor changes the one
large change is to make ReleaseGraphicsResources in Renderer
public like it is for every other class. This is the way it
should be and the original protected method was a mistake from
way way back.  Making it public save us from doing the

SetRenderWindow(NULL)
SetRenderWindow(this)

hack that we have been using to free up resources in the
renderer since we could nto call the protected method.

Change-Id: If31406fa9425417d240311aac110c82409637163