Bug #0015217: Fix tet aspect ratio computation.
AspectRatio is not correctly computed for tetrahedron that are not
well oriented regarding VTK specifications whereas TetAspectGamma
and TetAspectFrobenius do not care.
The patch, suggested by A.Ribes and A.Geay from EDF fixes the issue.
Also apply the patch for v_tet_aspect_beta, v_tet_collaspse_ratio,
and v_tet_condition.

Change-Id: I6a8ccd82634ffd46a45a07af2fbc7ac32cfa5e66