Fixed test to work on older graphics hardware
Older graphics hardware may support only textures with dimensions
that are powers of two. The TestActor2D test was running into the
behavior where the size of the texture used to store a vtkLookupTable
was being bumped up to the next power of two on hardware that supports
only power-of-two-sized textures. Hence, on older hardware, this test
could fail because the textured plane would wind up with too many
colors. This patch is a workaround that sets the number of colors in
the lookup table to a power of two minus 2 (the 2 missing colors in
the texture are filled by the out-of-range special colors). Hence,
the colors will fit in a texture whose size is a power of two.

The ideal fix would store only the colors defined in the
vtkLookupTable in the texture and pass the maximum valid
texture coordinate to the various parts of VTK that transform
data arrays to texture coordinates. Because modern graphics hardware
does not impose the power-of-two texture size limit, this
is not worth the trouble.

Change-Id: Ie89cf128f21ea65043ff68869ea18564a349fe37